Fixes to Criterias & Data

+++ Louvre-Cartes/Data/Criterias/DateCriteria.cs
@@ -10,7 +10,7 @@ namespace LouvreCartes.Data
     {
         protected override bool CheckCriteria(Mission mission, int count, Card card)
         {
-            bool inferior = mission.X is int value && value == 0;
+            bool inferior = mission.X is Int64 value && value == 0;
             return inferior ? (card.Date <= 0) : (card.Date >= 0);
         }
     }
